Permet desar hores jugades amb decimals (mitges hores)
hours_played era un enter tant a la BD com al model UserGame, així que
no es podien desar 12.5 hores, per exemple. Migra la columna a
numeric(6,1) i el camp a double a tota l'app; el camp d'hores accepta
ara punt o coma com a separador decimal.

+++ lib/repositories/supabase_library_repository.dart
@@ -106,7 +106,7 @@ class SupabaseLibraryRepository implements LibraryRepository {
     DateTime? completedAt,
     DateTime? droppedAt,
     DateTime? pausedAt,
-    int hoursPlayed = 0,
+    double hoursPlayed = 0,
     bool favorite = false,
   }) async {
     final user = client.auth.currentUser;
